requirements/base.txt
# AI System - Base Dependencies (~120MB)
# Core runtime dependencies for all containers

# Configuration & Environment
PyYAML>=6.0
python-dotenv>=1.0.0
click>=8.1.0
typer>=0.9.0

# Data Processing & Serialization  
orjson>=3.8.0
numpy>=1.24.0
pandas>=2.0.0
lz4>=4.0.0

# Network & Communication
pyzmq>=25.0.0
redis>=4.5.0
requests>=2.31.0
websockets>=11.0.0
asyncio-mqtt>=0.16.1

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
aiofiles>=23.0.0

# Validation & Models
pydantic>=2.0.0

# Monitoring & Observability
prometheus-client>=0.17.0
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0
structlog>=23.1.0

# System & Performance
psutil>=5.9.0
netifaces>=0.11.0
tenacity>=8.2.0
pybreaker>=1.0.0

# Utilities
rich>=13.4.0
tqdm>=4.65.0

# Testing (development)
pytest>=7.4.0
pytest-asyncio>=0.21.0 